[refactor]context 사용해서 페이지 분리
